small refactorings
