+++ pkg/thumbnails/resolutions/resolution_test.go
@@ -1,4 +1,4 @@
-package resolution
+package resolutions
 
 import "testing"
