+++ pkg/thumbnails/storage/storage.go
@@ -1,12 +1,12 @@
 package storage
 
-import "github.com/owncloud/ocis-thumbnails/pkg/thumbnails/resolution"
+import "github.com/owncloud/ocis-thumbnails/pkg/thumbnails/resolutions"
 
 // Context combines different attributes needed for storage operations.
 type Context struct {
 	ETag       string
 	Types      []string
-	Resolution resolution.Resolution
+	Resolution resolutions.Resolution
 }
 
 // Storage defines the interface for a thumbnail store.
